DSFAAP-2151: moving customers and srabpi plants to alpha types
