+++ src/types/alpha/customers.js
@@ -1,8 +1,8 @@
 import Joi from 'joi'
 
-import { LinksReference } from './links.js'
+import { LinksReference } from '../links.js'
 import { SrabpiPlantReference } from './srabpi-plants.js'
-import { Address } from './alpha/locations.js'
+import { Address } from './locations.js'
 
 const ContactDetails = Joi.object({
   fullName: Joi.string().allow(null, '').label('Full name'),
